% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AIC.gkwreg2}
\alias{AIC.gkwreg2}
\title{Akaike Information Criterion for gkwreg2 Object}
\usage{
\method{AIC}{gkwreg2}(object, ..., k = 2)
}
\arguments{
\item{object}{A fitted \code{gkwreg2} object.}

\item{...}{Additional \code{gkwreg2} objects for comparison.}

\item{k}{Penalty parameter (default 2 for AIC).}
}
\value{
AIC value(s).
}
\description{
Akaike Information Criterion for gkwreg2 Object
}
