% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.gkwreg2}
\alias{anova.gkwreg2}
\title{ANOVA for gkwreg2 Objects}
\usage{
\method{anova}{gkwreg2}(object, ..., test = c("Chisq", "LRT", "none"))
}
\arguments{
\item{object}{A fitted \code{gkwreg2} object.}

\item{...}{Additional \code{gkwreg2} objects to compare.}

\item{test}{Type of test: \code{"Chisq"} or \code{"LRT"} (equivalent),
or \code{"none"} to suppress p-values.}
}
\value{
A data frame of class \code{"anova"} with columns:
\describe{
\item{Model}{Model identifier}
\item{Df}{Degrees of freedom (number of parameters)}
\item{LogLik}{Log-likelihood}
\item{AIC}{Akaike Information Criterion}
\item{BIC}{Bayesian Information Criterion}
\item{Chisq}{Likelihood ratio test statistic (vs previous model)}
\item{Df diff}{Difference in degrees of freedom}
\item{Pr(>Chisq)}{P-value from chi-squared distribution}
}
}
\description{
Performs likelihood ratio tests to compare nested \code{gkwreg2} models.
}
\details{
Models should be nested for valid likelihood ratio tests. The function
compares each model to the previous one in the list, computing:
\deqn{LRT = -2(\ell_0 - \ell_1)}
where \eqn{\ell_0} is the log-likelihood of the simpler model.
}
\examples{
\donttest{
# Generate test data
set.seed(123)
n <- 150
x <- rnorm(n)
y <- gkwdist::rkw(n, alpha = exp(0.5 + 0.2 * x), beta = exp(1))
df <- data.frame(y = y, x = x)

# Compare nested models
fit0 <- gkwreg2(y ~ 1, data = df, family = "kw")
fit1 <- gkwreg2(y ~ x, data = df, family = "kw")
fit2 <- gkwreg2(y ~ x | x, data = df, family = "kw")

anova(fit0, fit1, fit2)
}

}
