% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_invlink_r}
\alias{apply_invlink_r}
\title{Apply Inverse Link Function (R version)}
\usage{
apply_invlink_r(eta, link_type)
}
\arguments{
\item{eta}{Linear predictor vector.}

\item{link_type}{Link function name.}
}
\value{
Transformed parameter values.
}
\description{
R wrapper for applying inverse link functions.
}
\keyword{internal}
