% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_link}
\alias{apply_link}
\title{Apply Link Function (forward transform)}
\usage{
apply_link(val, link_type)
}
\arguments{
\item{val}{Parameter value (in natural scale).}

\item{link_type}{Link function name.}
}
\value{
Transformed value (eta).
}
\description{
Transforms parameter value to linear predictor scale.
}
\keyword{internal}
