% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_functions.R
\name{apply_link_r}
\alias{apply_link_r}
\title{Apply Link Function (R wrapper)}
\usage{
apply_link_r(theta, link_type)
}
\arguments{
\item{theta}{Parameter values}

\item{link_type}{Link function name}
}
\value{
Transformed values (linear predictor scale)
}
\description{
R wrappers for applying link functions to parameter values.
}
\keyword{internal}
