% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_functions.R
\name{available_links}
\alias{available_links}
\title{Available Link Functions}
\usage{
available_links(type = c("positive", "probability", "all"))
}
\arguments{
\item{type}{Type of parameter: \code{"positive"} for positive parameters
(\eqn{\alpha}, \eqn{\beta}, \eqn{\gamma}, \eqn{\lambda}), or \code{"probability"} for probability parameters (\eqn{\delta} in (0,1)).}
}
\value{
Character vector of link function names.
}
\description{
Returns a character vector of available link function names.
}
\examples{
available_links("positive")
available_links("probability")

}
