% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{BIC.gkwreg2}
\alias{BIC.gkwreg2}
\title{Bayesian Information Criterion for gkwreg2 Object}
\usage{
\method{BIC}{gkwreg2}(object, ...)
}
\arguments{
\item{object}{A fitted \code{gkwreg2} object.}

\item{...}{Additional \code{gkwreg2} objects for comparison.}
}
\value{
BIC value(s).
}
\description{
Bayesian Information Criterion for gkwreg2 Object
}
