% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compute_mean_numerical}
\alias{compute_mean_numerical}
\title{Compute Mean via Numerical Integration}
\usage{
compute_mean_numerical(par, family)
}
\arguments{
\item{par}{Parameter vector for a single observation.}

\item{family}{Distribution family.}
}
\value{
Expected value E(Y).
}
\description{
Compute Mean via Numerical Integration
}
\keyword{internal}
