% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_parser.R
\name{extract_response}
\alias{extract_response}
\title{Extract Response Variable from Formula}
\usage{
extract_response(formula, data)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{Data frame.}
}
\value{
Numeric vector of response values.
}
\description{
Extract Response Variable from Formula
}
\keyword{internal}
