% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{formula.gkwreg2}
\alias{formula.gkwreg2}
\title{Extract Formula from gkwreg2 Object}
\usage{
\method{formula}{gkwreg2}(x, ...)
}
\arguments{
\item{x}{A fitted \code{gkwreg2} object.}

\item{...}{Currently ignored.}
}
\value{
The model formula.
}
\description{
Extract Formula from gkwreg2 Object
}
