% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_default_links}
\alias{get_default_links}
\title{Get Default Link Functions for Family}
\usage{
get_default_links(family)
}
\arguments{
\item{family}{Distribution family name.}
}
\value{
Named list of link function names.
}
\description{
Returns the default link functions for each parameter of a distribution family.
All parameters are positive (> 0), so log link is appropriate.
}
\details{
Parameter spaces for each family:
\itemize{
\item kw: alpha > 0, beta > 0
\item beta: gamma > 0, delta > 0 (shape1, shape2)
\item ekw: alpha > 0, beta > 0, lambda > 0
\item mc: gamma > 0, delta > 0, lambda > 0
\item bkw: alpha > 0, beta > 0, gamma > 0, delta > 0
\item kkw: alpha > 0, beta > 0, delta > 0, lambda > 0
\item gkw: alpha > 0, beta > 0, gamma > 0, delta > 0, lambda > 0
}

All parameters are strictly positive, so log link is the natural choice
as it maps R -> (0, inf).
}
\examples{
get_default_links("kw")
# $alpha: "log", $beta: "log"

}
