% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.gkwreg2}
\alias{logLik.gkwreg2}
\title{Extract Log-Likelihood from gkwreg2 Object}
\usage{
\method{logLik}{gkwreg2}(object, ...)
}
\arguments{
\item{object}{A fitted \code{gkwreg2} object.}

\item{...}{Currently ignored.}
}
\value{
A \code{logLik} object with attributes \code{df} (degrees of freedom)
and \code{nobs} (number of observations).
}
\description{
Extract Log-Likelihood from gkwreg2 Object
}
