% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{model.frame.gkwreg2}
\alias{model.frame.gkwreg2}
\title{Model Frame for gkwreg2 Object}
\usage{
\method{model.frame}{gkwreg2}(formula, ...)
}
\arguments{
\item{formula}{A fitted \code{gkwreg2} object.}

\item{...}{Currently ignored.}
}
\value{
The model frame used for fitting.
}
\description{
Model Frame for gkwreg2 Object
}
