% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{model.matrix.gkwreg2}
\alias{model.matrix.gkwreg2}
\title{Extract Model Matrix from gkwreg2 Object}
\usage{
\method{model.matrix}{gkwreg2}(object, parameter = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{gkwreg2} object.}

\item{parameter}{Which parameter's design matrix to extract.}

\item{...}{Currently ignored.}
}
\value{
A design matrix, or list of matrices if \code{parameter} not specified.
}
\description{
Extract Model Matrix from gkwreg2 Object
}
