% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.gkwreg2}
\alias{nobs.gkwreg2}
\title{Number of Observations for gkwreg2 Object}
\usage{
\method{nobs}{gkwreg2}(object, ...)
}
\arguments{
\item{object}{A fitted \code{gkwreg2} object.}

\item{...}{Currently ignored.}
}
\value{
Number of observations used in fitting.
}
\description{
Number of Observations for gkwreg2 Object
}
