% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{objective_gkwreg}
\alias{objective_gkwreg}
\title{Compute negative log-likelihood for gkwreg2}
\usage{
objective_gkwreg(
  theta_concat,
  y,
  X_list,
  family,
  link_list,
  param_indices,
  n_coefs
)
}
\arguments{
\item{theta_concat}{Concatenated regression coefficients}

\item{y}{Response vector (values in (0,1))}

\item{X_list}{List of design matrices (one per parameter)}

\item{family}{Distribution family name}

\item{link_list}{Named list of link functions per parameter}

\item{param_indices}{Integer vector with start indices for each parameter}

\item{n_coefs}{Integer vector with number of coefficients per parameter}
}
\value{
Negative log-likelihood value
}
\description{
Compute negative log-likelihood for gkwreg2
}
\keyword{internal}
