% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{objective_gkwreg_vec}
\alias{objective_gkwreg_vec}
\title{Compute negative log-likelihood using vectorized gkwdist calls}
\usage{
objective_gkwreg_vec(
  theta_concat,
  y,
  X_list,
  family,
  link_list,
  param_indices,
  n_coefs
)
}
\arguments{
\item{theta_concat}{Concatenated regression coefficients}

\item{y}{Response vector}

\item{X_list}{List of design matrices}

\item{family}{Distribution family}

\item{link_list}{Named list of link functions}

\item{param_indices}{Start indices}

\item{n_coefs}{Number of coefficients per parameter}
}
\value{
Negative log-likelihood value
}
\description{
Compute negative log-likelihood using vectorized gkwdist calls
}
\keyword{internal}
