% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{print.anova.gkwreg2}
\alias{print.anova.gkwreg2}
\title{Print ANOVA Table}
\usage{
\method{print}{anova.gkwreg2}(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An \code{anova.gkwreg2} object.}

\item{digits}{Number of significant digits.}

\item{signif.stars}{Logical: print significance stars?}

\item{...}{Currently ignored.}
}
\description{
Print ANOVA Table
}
