% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{print.gkw_control}
\alias{print.gkw_control}
\title{Print gkw_control Object}
\usage{
\method{print}{gkw_control}(x, ...)
}
\arguments{
\item{x}{A \code{gkw_control} object.}

\item{...}{Ignored.}
}
\description{
Print gkw_control Object
}
