% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.gkwreg2}
\alias{print.gkwreg2}
\title{Print gkwreg2 Object}
\usage{
\method{print}{gkwreg2}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A fitted \code{gkwreg2} object.}

\item{digits}{Number of significant digits for printing.}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Displays a concise summary of a fitted \code{gkwreg2} model.
}
