% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.gkwreg2}
\alias{print.summary.gkwreg2}
\title{Print Summary of gkwreg2 Object}
\usage{
\method{print}{summary.gkwreg2}(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{A \code{summary.gkwreg2} object.}

\item{digits}{Number of significant digits.}

\item{signif.stars}{Logical: print significance stars?}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Summary of gkwreg2 Object
}
