% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{update.gkwreg2}
\alias{update.gkwreg2}
\title{Update a gkwreg2 Model}
\usage{
\method{update}{gkwreg2}(object, formula., ..., evaluate = TRUE)
}
\arguments{
\item{object}{A fitted \code{gkwreg2} object.}

\item{formula.}{Changes to the formula.}

\item{...}{Additional arguments to \code{gkwreg2}.}

\item{evaluate}{If \code{TRUE}, evaluate the updated call.}
}
\value{
Updated \code{gkwreg2} object or call.
}
\description{
Update a gkwreg2 Model
}
