% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{gkwreg2-package}
\alias{gkwreg2-package}
\title{gkwreg2: Generalized Kumaraswamy Regression Models}
\description{
The gkwreg2 package provides regression modeling for bounded data in (0,1)
using the Generalized Kumaraswamy distribution family. It implements
maximum likelihood estimation with analytical gradients via RcppArmadillo.
}
\section{Distribution Families}{

\itemize{
\item \strong{gkw}: Generalized Kumaraswamy (5 parameters: \eqn{\alpha}, \eqn{\beta}, \eqn{\gamma}, \eqn{\delta}, \eqn{\lambda})
\item \strong{bkw}: Beta-Kumaraswamy (4 parameters: \eqn{\alpha}, \eqn{\beta}, \eqn{\gamma}, \eqn{\delta})
\item \strong{kkw}: Kumaraswamy-Kumaraswamy (4 parameters: \eqn{\alpha}, \eqn{\beta}, \eqn{\delta}, \eqn{\lambda})
\item \strong{ekw}: Exponentiated Kumaraswamy (3 parameters: \eqn{\alpha}, \eqn{\beta}, \eqn{\lambda})
\item \strong{mc}: McDonald (3 parameters: \eqn{\gamma}, \eqn{\delta}, \eqn{\lambda})
\item \strong{kw}: Kumaraswamy (2 parameters: \eqn{\alpha}, \eqn{\beta})
\item \strong{beta}: Beta (2 parameters: \eqn{\gamma}, \eqn{\delta})
}
}

\section{Main Functions}{

\itemize{
\item \code{\link{gkwreg2}}: Main model fitting function
\item \code{\link{gkw_control}}: Control parameters for optimization
\item \code{\link{get_param_names}}: Get parameter names for a family
\item \code{\link{get_default_links}}: Get default link functions
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/evandeilton/gkwreg2}
  \item \url{https://evandeilton.github.io/gkwreg2/}
  \item Report bugs at \url{https://github.com/evandeilton/gkwreg2/issues}
}

}
\author{
Evandeilton Lopes
}
